+++ src/main/java/kitchenpos/order/domain/JpaOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
