+++ src/main/java/kitchenpos/order/ui/ProductRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.ui;
+package kitchenpos.order.ui;
 
 import kitchenpos.product.application.ProductService;
 import kitchenpos.product.domain.Product;
