+++ src/test/java/kitchenpos/Fixtures.java
@@ -1,9 +1,13 @@
 package kitchenpos;
 
-import kitchenpos.domain.*;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderType;
 import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
